requirements.txt
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.0
click==8.1.7
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
Django==4.2.5
django-cors-headers==4.2.0
django-filter==23.3
django-rest-swagger==2.2.0
django-tinymce==3.7.1
Django-Verify-Email==2.0.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
drf-yasg==1.21.7
h11==0.14.0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.3
openapi-codec==1.3.2
packaging==23.1
Pillow==10.0.0
PyJWT==2.8.0
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
simplejson==3.19.2
sqlparse==0.4.4
typing_extensions==4.7.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.0.6
uvicorn==0.29.0
whitenoise==6.6.0
